backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain==0.1.0
langchain-community==0.0.16
langchain-core==0.1.18
chromadb==0.4.22
sentence-transformers==2.2.2
pypdf2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
python-dotenv==1.0.0
numpy==1.24.3
Pillow==10.2.0 